Reverted the revert on tcpsocket branch.
Sorry for any confusion here, I made mistakes when trying to revert from main.
My previous two commits are titled the same, but the second reverts the first. This commit will revert the second so my changes will end up saved in tcpsocket branch.

+++ cli_user_functions.py
@@ -157,7 +157,7 @@ if __name__ == "__main__":
     if len(sys.argv) < 2:
         sys.exit(1)
     
-    port = sys.argv[1]
+    port = (int)(sys.argv[1])
     event = threading.Event()
     cli_thread = threading.Thread(target=shell_loop, args=(port, event))
     cli_thread.start()
